+++ tests/apps/aiohttp_app.py
@@ -1,4 +1,3 @@
-import json
 import asyncio
 
 import aiohttp
@@ -8,7 +7,7 @@ import marshmallow as ma
 
 from webargs import fields
 from webargs.aiohttpparser import parser, use_args, use_kwargs
-from webargs.core import MARSHMALLOW_VERSION_INFO
+from webargs.core import MARSHMALLOW_VERSION_INFO, json
 
 hello_args = {"name": fields.Str(missing="World", validate=lambda n: len(n) >= 3)}
 hello_multiple = {"name": fields.List(fields.Str())}